requirements.txt
numpy==1.23.3
#numba==0.48.0
torch>=1.1
tensorboardX
easydict
pyyaml
scikit-image
tqdm
torchvision
SharedArray
opencv-python
pyquaternion
spconv-cu113
av2
kornia
torch_scatter
shapely==1.8.4
llvmlite
